#deploy: Push notification module and many many new features

+++ src/main/java/mn/usug/dis_news_service/DTO/VehicleOrderItemSaveDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class VehicleOrderItemSaveDto {
 
-    private Long vehicleTypeId;
+    private Integer vehicleTypeId;
 
     // Ердийн машинд
     private Integer qty;
